Implementação do DataReceiver
